// src/app/core-module/enums/field-name.enum.ts
export enum FieldNameEnum {
  // user details component
  FIRST_NAME = 'firstName',
  LAST_NAME = 'lastName',
  EMAIL = 'email',
  PHONE_NUMBER = 'phoneNumber',
  PASSWORD = 'password',
  NICK_NAME = 'nickName',
  UNIVERSITY_NAME = 'universityName',
  MAJOR = 'major',
  INTAKE_ID = 'intakeId',
  YEAR = 'year',
  // common

  COMMENTS = 'comments',

  // interview details component
  INTERVIEW_DATE = "interviewDate",
  TRAVEL_DATE = "travelDate"
}